--- src/main/java/Hammurabi.java
@@ -1,56 +1,0 @@
-import java.util.Random;
-import java.util.Scanner;
-
-public class Hammurabi {
-    Random rand;
-    Scanner scanner=new Scanner(System.in);
-
-    public static void main ( String[] args ) {
-        new Hammurabi().playGame();
-
-    }
-
-    private void playGame () {   //each year upto 10 years
-        int population;
-        int grain;
-        int acres;
-        //int value;
-
-    }
-
-    public int askHowManyAcresToBuy(int price,int bushels){
-        return 0;
-    }
-
-
-    public int askHowManyAcresToSell(int acresOwned){
-        return 0;
-    }
-
-    public  int askHowManyAcresToPlant(int acresOwned,int population,int bushels){
-        return 0;
-    }
-
-    public int plagueDeaths(int population){
-        return 0;
-    }
-
-    public boolean uprising(int population,int howManyPeopleStarved){
-        return false;
-    }
-    int immigrants(int population, int acresOwned, int grainInStorage){
-        return 0;
-    }
-    int harvest(int acres, int bushelsUsedAsSeed){
-        return 0;
-    }
-    int grainEatenByRats(int bushels){
-        return 0;
-    }
-
-    int newCostOfLand(){
-        return 0;
-    }
-
-
-}
